Move About to header, remove purple boxes
Removes the purple degradations at header and footer, doing away with
the later altogether.

I should probably align the margins of the top of the header (name,
professional title, contact info) with the rest of the document. But I
need that extra space when I swap my data in.

I may try decreasing font size for the contact info flex.