Overhaul the header elements
Adds support for displaying a headshot on the same row as name + title,
right aligned. To achieve that, refactors a lot of existing code.

Also makes the contact-info flexbox space-between justified, so it takes
up the whole width of the document.

There's still some work left to polish the header (see the //todo
markers) but should be easy enough.