Make data.yml actually YAML
By importing the raw file with Webpack's raw-loader.

The purpose of this is so that the linter/syntax highlighting in vscode
doesn't sound alarms all throughout the data file, while I'm editing it.

Reference for this solution:
https://chrisfrew.in/blog/import-a-file-as-a-raw-string-in-gatsby-or-react/

Also, raw-loader is deprecated for Webpack v5 (we're v4). If we update,
it'd be ideal to drop this dependency and use the built-in asset/source:
https://webpack.js.org/guides/asset-modules/