Fix Chrome issues
I was testing on Firefox and exporting to PDF not as often as I
should have. The project uses Puppeteer (Chrome-based) to render and
export to PDF. So when I checked, it turns out that quite a few things
didn't look very good.

In particular, I needed to import the 600 weight version of the Barlow
font because Chrome was using a very chunky version of the font for the
headers.

Aside from that, some side improvements:

- made the top header (full name + profession) leaner by decreasing
font-weight from 600 to 500.
- made the job-bullets align their text left, like in the Parker
template I'm using as reference. It looks much better (ideally with
bullet points <3 lines).