Add toggle to fix Chromium blurry photo rendering
This is Chrome/ium-specific, it blurs the image for some reason. It only
happens in dev, as the pdf export with Puppeteer looks fine, same as in
Firefox.

This commented CSS is picked up by Chrome DevTools, and then it can be
toggled on and off to preview how the photo will actually look in the
export.

The picture doesn't seem to render as well as in FF/export, but it's
still much better than the blurry default.

Relevant StackOverflow: https://superuser.com/a/530382