Fix skill separators
Problem: skill separators (implemented with border-left) looked well
enough when opening the generated pdf in Firefox or Evince, but awfully
thick on Chromium. Thew were blocks rather than I-beam.

It's shitty because they look well enough in the Chromium live preview
(when you `$ npm run dev` the server to localhost).

I tried a bunch of things before this fix, but none worked. Namely:
- changing border width (subpixel, thin)
- using box-shadow rather than border to draw the separators
- updating puppeteer (and so the Chrome/Skia version)
- printing from Chromium instead of puppeteer (pretty much the same)
- different combinations of border, shadow-box and outline, some white
- replacing the flexbox for floats
- changing ul/li to divs
- etc.

At some point, on the version that used box-shadows with a trick to draw
them on one side only [1], I noticed that the printed document had
irregular shadows. Some separators where thicker than others.

Then I searched how to fix that and found discussions about this in [2],
[3]. The solution is setting `-webkit-print-color-adjust: exact`
(already in the `src/pages/resume.vue` wrapper), and also
`webkit-filter: opacity(1)`.

Problem is, setting `-webkit-filter` rasterizes the element it's set on.
That makes its text un-parseable, which is a no-no considering automated
resume screening.

So the same SO discussion in [2] contains a refined version of the
solution. [4] That's the one I used. And it works with border too, so I
rolled shadow-box back.

Lastly, I also added some margins to the separator pseudo-element to
make it shorter, cause the border as-is looks too tall next to the skill
names, unnecessarily draws attention.

[1] https://css-tricks.com/almanac/properties/b/box-shadow/#one-side-only
[2] https://stackoverflow.com/q/13975198
[3] https://github.com/puppeteer/puppeteer/issues/5284
[4] https://stackoverflow.com/a/57714709