Improve font rendering
Tested on Chromium, Firefox and Evince. In general it makes the text
look more compact and consistent, as if the previous version used
justified text. On Evince, it also fixes an ugly overlap between some
letter combinations, and a large distance between some others.

Discovered when I was printing from Chromium to quicken development. The
browser-printed version had different font width and lines wrapped on
different words. Then I found discussions on this on [1, 2].

From the solutions suggested in [2], both `--font-render-hinting=medium`
and `--font-render-hinting=none` work well, they look exactly the same.
I went with the former because it sounds less drastic.

I've only tested for the Marshall template but it should benefit other
templates too.

[1] https://github.com/puppeteer/puppeteer/issues/4177
[2] https://github.com/puppeteer/puppeteer/issues/2410