Style the no-photo version
I've considered aligning the contact-info icons to the right, but it'd
be a hassle: they're on the left in marshal-with-photo, I have no idea
how I'd handle that without duplicating a lot of code or breaking it
into a separate component.

And it wouldn't even look good, cause icons have different widths and
aligning them looks ugly.

The transition with/out photo looks is neat with my data too, low
maintainance, because the height of 5 contact-info rows is more or less
the same height as the photo it replaces.